refactor(slot): Move SlotInteractionResult to internal package
