+++ core/src/main/kotlin/CustomInventory.kt
@@ -33,9 +33,9 @@ import org.bukkit.inventory.ItemStack
 import ru.endlesscode.inventory.internal.TaskScheduler
 import ru.endlesscode.inventory.internal.di.DI
 import ru.endlesscode.inventory.internal.listener.*
+import ru.endlesscode.inventory.internal.listener.SlotInteractionResult.Change
 import ru.endlesscode.inventory.internal.util.*
 import ru.endlesscode.inventory.slot.*
-import ru.endlesscode.inventory.slot.SlotInteractionResult.Change
 import java.util.*
 
 /**
